AM-450 : Refactoring and minor updates.
