+++ src/main/java/uk/gov/hmcts/ccd/v2/external/controller/CaseDocumentController.java
@@ -4,7 +4,6 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -23,7 +22,7 @@ public class CaseDocumentController {
     private final GetCaseDocumentOperation getCaseDocumentOperation;
 
     @Autowired
-    public CaseDocumentController(@Qualifier("default") GetCaseDocumentOperation getCaseDocumentOperation) {
+    public CaseDocumentController(GetCaseDocumentOperation getCaseDocumentOperation) {
         this.getCaseDocumentOperation = getCaseDocumentOperation;
     }
 
@@ -66,7 +65,7 @@ public class CaseDocumentController {
     public ResponseEntity<CaseDocumentResource> getCaseDocumentMetadata(@PathVariable("caseId") String caseId,
                                                                         @PathVariable("documentId") String documentId) {
 
-        final CaseDocumentMetadata documentMetadata = this.getCaseDocumentOperation.execute(caseId,documentId);
+        final CaseDocumentMetadata documentMetadata = this.getCaseDocumentOperation.getCaseDocumentMetadata(caseId,documentId);
         return ResponseEntity.ok(new CaseDocumentResource(caseId, documentId, documentMetadata));
     }
 }
